Calculate differential expression statistics by a grouping variable.

mt_diffexprs

mt_diffexprs(obj,group)

Arguments

obj
(required) An object of class 'mmt' (see mt_load) or 'ampvis2' (see amp_load).

group
(required) Character string defining the grouping variable to use for differential expression.

intercept
(optional) Specify the level of 'group' to be used as reference/intercept. Default is by alphabetical order.

row_labels
(optional) Specify columns of rowdata to print on heatmap.

order_var_by
(optional) How to order rows of heatmap. You can specify 1) "variance" or 2) one of the levels in group, which will order according to the absolute value of that level. (default: "variance")

row_show
(optional) The number of rows to show.

Value

A list with 2 elements:

heatmap - A heatmap showing the log fold-change of each level in group relative to the intercept/baseline level specified in intercept. The intercept is therefore not possible to plot.

DESeq - A DESeqDataSet containing the results of the differential expression analysis. To extract results and do comparisons between groups use the mt_dumpDE function.

Details

Typically one would subset low-expression genes (use mt_subset) before performing the analysis, due to a low signal/noise ratio. This also drastically reduces computation times and improves stability of the parameter estimation.

Examples

# NOT RUN {
# Load data.
data("example_mmt")

# Compute the statistics.
DE_mt <- mt_diffexprs(example_mmt,
  group      = "Type",
  row_labels = c("GeneID","product"),
  intercept  = "ANAMMOX")

# Extract results for given levels.
mt_res <- mt_dumpDE(DE_mt,nom = "ELECTRODE",denom = "SUSPENTION")

# Show the output.
mt_res$MAplot
mt_res$BOXplot
head(mt_res$Table)
# }