Make a stacked barplot of the sequencing statistics.

mt_stackbar

mt_stackbar(mmt,Raw,after_QC)

Arguments

mmt
(required) An object of class mmt.

group_by
(optional) A column name in the metadata to group samples by. (Default: NULL)

round_to
(optional) Round the counts to thousands ("K") or milions ("M"). (Default: "M")

size_axis
(optional) Specify size of primary and secondary axis. (Default: 10)

size_text
(optional) Specify size of text in plot. (Default: 3)

detailed
(optional) Output a table with the data used for the plot.

Value

A ggplot

Details

The mt_stackbar function only works on data generated from the MetaT pipeline and loaded using the mt_loadMetaT function. It is meant for data quality assesment by giving a quick overview of the proportions of data removed for each pre-processing step, if this is available in the metadata.

Examples

# NOT RUN {
data("example_mmt")

mt_stackbar(example_mmt,Raw = "Raw",after_QC = "after_QC",after_filter = "after_filter")
# }